requirements.txt
baidu-aip==2.2.17.0
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
idna==2.8
imageio==2.6.1
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.0.3
networkx==2.4
numpy==1.16.4
opencv-contrib-python==3.4.2.17
pandas==0.24.2
Pillow==6.1.0
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
requests==2.22.0
Rtree==0.8.3
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.2.1
six==1.13.0
tesserocr==2.4.0
urllib3==1.25.7
torch==1.1.0

